Create a launcher script to automatically select the updater for Intel or ARM
Instead of having eos-chrome-updater for Intel and eos-chrome-updater-arm for
ARM, have an eos-chrome-updater-intel script for Intel and let the unsuffixed
one to be just a launcher that will select which one to launch based on the
current kernel architecture.

This makes things more consistent, and makes it easier for users / ground team
to run the updater, as there will be one single entry point.

https://phabricator.endlessm.com/T5706