Disable integration with Network Manager for i386 as well now
Since March 2nd, Google no longer publishes 32-bit binary builds of
Google Chrome, so we disable the automatic hook on if-up events as
that the uploader script will no longer work in that architecture.

Still this package is still useful for current installations as it
provides the necessary symlink for the Widevine plugin to work and
the runtime flags for chromium to find the Flash plugin, in case
they are present under /var/lib/chromium-plugins-extra.

https://phabricator.endlessm.com/T10669