Don't make Widevine installation dependand on the Flash plugin
Currently, if the Flash plugin was not available for some reason, the
whole installation would abort leaving us without the Widevine plugin
even when it was there (meaning "no Netflix", for instance).

This, while not a common situation, can happen as it did now due
to a change in how the Flash plugin is distributed, and so we should
not make one plugin's existence dependant on other's.

https://phabricator.endlessm.com/T13777