Add example file describing the Last Known Good Version of the flash plugin
This type of files will be useful to decouple this from OSTree upgrades,
since it allows us to update the last version just by updating this file
in the repo.

Note that this is just a example file to show how its contents would be
for the particular case of the 23.0.0.185 version, but the real location
the script will check for will be served from one of our S3 buckets.

https://phabricator.endlessm.com/T13777