Refactor upstream chrome plugin updater in preparation for Flash
Until now, we used to retrieve both the Widevine and Flash plugins via
the Google Chrome debian package, so most of the steps were shared and
a failure in one of those steps would normally mean problems for both
plugins, not just for one.

From now on, we will be downloading Flash from Adobe's own service but
retrieving the Widevine plugin from Chrome's debian package, so we need
to reorganize the logic here a bit so that we can decouple the Widevine
bits from the Flash bits as much as possible, while sharing the same unit.

This mainly means moving away from early returns (e.g. fp_exit_with_error)
in case something goes wrong when downloading Widevine (as we might still
have a chance with Flash now), and reorganizing the code a bit too.

https://phabricator.endlessm.com/T13777