Add logic to download and install the Flash plugin from Adobe's servers
Similar to what it's done for Widevine, we download the tarball from
Adobe's servers and install it in its right destination, updating the
VERSION file (now Flash_VERSION.txt) as needed, so that we can later on
check whether we even need to check for new updates.

As hinted in the previous commit, the Last Known Good Version of the
flash plugin will be retrieved from a file in our git repository,
falling back to the details of a specific version (the last one at
this moment) in case that file can't be retrieved.

https://phabricator.endlessm.com/T13777