avoid plugin download when local firewall is active
This is for certain deployments which are known to have internet
connectivity, but it is very constrained, such that having all of the
computers attempting to download Chrome & various plugins would consume
an undesirably high amount of bandwidth. These systems have a local
firewall which prevents general access to the internet. To avoid logging
confusing/distracting errors, gracefully disable this script when we know
the firewall will cause it to fail.

Note that as well as disabling this script in the case that the firewall
is running, we also disable it in the case that the firewall is enabled -
we don't want to automatically download the plugin during some unrelated
administration task if we're also not confident that normal connectivity
will be available in future to keep it up to date.

https://phabricator.endlessm.com/T15547