add a list of products where the automatic download is disabled
This is for certain deployments which are known to have internet
connectivity, but it is very constrained, such that having all of
the computers attempting to download Chrome & various plugins would
consume an undesirably high amount of bandwidth. Unless/until a
more bandwidth-efficient solution is available, disable the
download on these images by matching the product name. This allows
the change to be reverted at some point in the future by updating
the script. (Disabling the script / systemd jobs in configuration
would be far harder to reverse over the air.)