update-upstream: Update url to fetch latest widevine version
The current URL (https://dl.google.com/widevine-cdm/current.txt) used
to fetch the latest widevine version is outdated and the widevine version
it points to doesn't work with chromium >= 77.

Google also provides an alternative URL (
https://dl.google.com/widevine-cdm/versions.txt) listing available
versions that is currently getting updated and with the latest version
working with chromium 77 and 78, so lets use that instead.

https://phabricator.endlessm.com/T27973